Move reviewCode functionality into vortex
